Move cp_benchmark chart to another file
